requirements.txt
dacite==1.9.2
pyyaml==6.0.2
requests==2.32.3
